@@ -13,7 +13,8 @@
     "eslint-fix": "eslint --ext ts,tsx --max-warnings 0 --fix src test example",
     "eslint": "eslint --ext ts,tsx --max-warnings 0 src test example",
     "prepare": "husky install",
-    "pretty-quick-staged": "pretty-quick --staged"
+    "pretty-quick-staged": "pretty-quick --staged",
+    "upgrade": "npm-check-updates --upgrade"
   },
   "keywords": [
     "gherkin",
@@ -24,37 +25,36 @@
   "author": "Aslak Hellesøy",
   "license": "MIT",
   "dependencies": {
-    "@cucumber/cucumber-expressions": "^13.0.1",
+    "@cucumber/cucumber-expressions": "^15.2.0",
     "@cucumber/gherkin-utils": "^7.0.0",
-    "@cucumber/language-service": "^0.2.0",
-    "@cucumber/messages": "^17.1.0",
-    "@cucumber/suggest": "^0.0.4",
-    "autocompleter": "^6.1.2",
-    "blockly": "^6.20210701.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@cucumber/language-service": "^0.26.0",
+    "@cucumber/messages": "^18.0.0",
+    "blockly": "^8.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   },
   "devDependencies": {
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.9.1",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "@typescript-eslint/eslint-plugin": "^4.31.0",
-    "@typescript-eslint/parser": "^4.31.0",
-    "esbuild": "^0.12.26",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
+    "esbuild": "^0.14.39",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.25.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "husky": "^7.0.2",
-    "mocha": "^9.1.1",
-    "prettier": "^2.4.0",
-    "pretty-quick": "^3.1.1",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.2"
+    "husky": "^8.0.1",
+    "mocha": "^10.0.0",
+    "npm-check-updates": "13.0.3",
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.3",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.7.2"
   },
   "prettier": {
     "trailingComma": "es5",
