@@ -4,57 +4,52 @@ import {
   ParameterTypeRegistry,
 } from '@cucumber/cucumber-expressions'
 import { pretty } from '@cucumber/gherkin-utils'
+import { buildSuggestions } from '@cucumber/language-service'
 import { GherkinDocument } from '@cucumber/messages'
-import { buildStepDocuments } from '@cucumber/suggest'
 import React, { useMemo, useState } from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 
 import BlocklyComponent from '../src/BlocklyComponent'
 import { defineBlocks } from '../src/defineBlocks.js'
 
-const toolbox = `
-<xml xmlns="https://developers.google.com/blockly/xml">
-  <block type="feature"/>
-  <block type="rule"/>
-  <block type="background"/>
-  <block type="scenario"/>
-  <block type="step"/>
-</xml>
-`
-
 // Build some sample step texts and cucumber expressions. These would typically come from a stream
 // of Cucumber Messages.
-const ef = new ExpressionFactory(new ParameterTypeRegistry())
+const registry = new ParameterTypeRegistry()
+const ef = new ExpressionFactory(registry)
 const expressions: Expression[] = [
-  ef.createExpression('I have {int} cukes in my belly'),
+  ef.createExpression('I have {int} cukes in my {word}'),
   ef.createExpression('there are {int} blind mice'),
 ]
-const stepDocuments = buildStepDocuments(
+const suggestions = buildSuggestions(
+  registry,
   ['I have 42 cukes in my belly', 'I have 96 cukes in my belly', 'there are 38 blind mice'],
   expressions
 )
-defineBlocks(stepDocuments)
+defineBlocks(suggestions)
 
 const App: React.FunctionComponent = () => {
   const [workspaceXml, setWorkspaceXml] = useState(`
     <xml xmlns="https://developers.google.com/blockly/xml">
-      <block type="scenario" id="H]D);F2kYo=xCdjp;[,}" x="78" y="-25">
-        <field name="SCENARIO_NAME">The one where...</field>
-        <statement name="STEPS">
-          <block type="step" id="7n3=OA!N#Yx.cyUYC^y5">
-            <field name="STEP_KEYWORD">GIVEN</field>
-            <field name="STEP_TEXT">this is an undefined step</field>
-            <next>
-              <block type="_cucumber_0" id="/fSm?XET8tbqrFUS24WH">
-                <field name="STEP_KEYWORD">WHEN</field>
-                <field name="STEP_FIELD_1">I have </field>
-                <field name="STEP_FIELD_2">8899</field>
-                <field name="STEP_FIELD_3"> cukes in my belly</field>
-              </block>
-            </next>
-          </block>
-        </statement>
-      </block>
+<!--      <block type="scenario" id="0V5*@yYI7-3LDlQ~%cP." x="76" y="101">-->
+<!--        <field name="SCENARIO_NAME">The one where...</field>-->
+<!--        <statement name="STEPS">-->
+<!--          <block type="I have {int} cukes in my {word}" id="f$CNPC;*bML~99P_\`prh">-->
+<!--            <field name="STEP_KEYWORD">GIVEN</field>-->
+<!--            <field name="STEP_FIELD_1">I have </field>-->
+<!--            <field name="STEP_FIELD_2">38</field>-->
+<!--            <field name="STEP_FIELD_3"> cukes in my belly</field>-->
+<!--            <next>-->
+<!--              <block type="there are {int} blind mice" id="@}nG:J{47KEaPV.uV2P\`">-->
+<!--                <field name="STEP_KEYWORD">GIVEN</field>-->
+<!--                <field name="STEP_FIELD_1">there are </field>-->
+<!--                <field name="STEP_FIELD_2">38</field>-->
+<!--                <field name="STEP_FIELD_3"> blind mice</field>-->
+<!--                <next></next>-->
+<!--              </block>-->
+<!--            </next>-->
+<!--          </block>-->
+<!--        </statement>-->
+<!--      </block>-->
     </xml>
   `)
   const [gherkinDocuments, setGherkinDocuments] = useState<readonly GherkinDocument[]>([])
@@ -67,6 +62,7 @@ const App: React.FunctionComponent = () => {
     <div className="flex-container">
       <div className="flex-child">
         <BlocklyComponent
+          suggestions={suggestions}
           workspaceXml={workspaceXml}
           setWorkspaceXml={setWorkspaceXml}
           setGherkinDocuments={setGherkinDocuments}
@@ -74,7 +70,6 @@ const App: React.FunctionComponent = () => {
             readOnly: false,
             trashcan: true,
             media: 'media/',
-            toolbox,
             move: {
               scrollbars: true,
               drag: true,
@@ -92,4 +87,7 @@ const App: React.FunctionComponent = () => {
   )
 }
 
-ReactDOM.render(<App />, document.getElementById('app'))
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const container = document.getElementById('app')!
+const root = createRoot(container)
+root.render(<App />)
